# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Invoice Generator API",
...
    "version": "1.0.0",
    "summary": "Professional Invoice Generator API",
    "contact": {
      "name": "Invoice Generator API Support Team",
      "email": "support@yourdomain.com",
      "url": "https://yourdomain.com/contact"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://yourdomain.com/terms"
  },
  "servers": [
    {
      "url": "https://screeching-tildi-adelzidoune-ca9a7151.koyeb.app",
      "description": "Koyeb Production Server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Create a new user account\n\n**No authentication required**\n\nCreates a new user with:\n- Unique email and username\n- Hashed password (bcrypt)\n- Optional company information\n- Automatic payment link generation",
        "operationId": "register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "username", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "freelancer@example.com",
                    "description": "Valid email address"
                  },
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "example": "freelancer",
                    "description": "Unique username (3-50 characters)"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "minLength": 8,
                    "example": "securepass123",
                    "description": "Strong password (min 8 characters)"
                  },
                  "full_name": {
                    "type": "string",
                    "example": "Ahmed Freelancer",
                    "description": "Full name (optional)"
                  },
                  "company_name": {
                    "type": "string",
                    "example": "Freelance Pro",
                    "description": "Company name (optional)"
                  },
                  "phone": {
                    "type": "string",
                    "example": "+212600000000",
                    "description": "Phone number (optional)"
                  },
                  "address": {
                    "type": "string",
                    "example": "123 Street, Casablanca, Morocco",
                    "description": "Full address (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "User created successfully"},
                    "user_id": {"type": "integer", "example": 1},
                    "username": {"type": "string", "example": "freelancer"}
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email or username already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Email already registered"}
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "loc": {"type": "array"},
                          "msg": {"type": "string"},
                          "type": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate user and return JWT token\n\n**No authentication required**\n\nReturns:\n- JWT access token (valid for 30 minutes)\n- Token type (Bearer)\n- User information",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "freelancer",
                    "description": "Username or email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securepass123",
                    "description": "User password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "token_type": {
                      "type": "string",
                      "example": "bearer"
                    },
                    "expires_in": {
                      "type": "integer",
                      "example": 1800,
                      "description": "Token expiration time in seconds"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {"type": "string", "example": "Invalid username or password"}
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme.\n\n**How to use:**\n1. Login at `/auth/login`\n2. Copy the `access_token` from response\n3. Click 🔓 **Authorize** button above\n4. Enter: `Bearer <your_token>`\n5. Click **Authorize**\n\n**Token lifetime:** 30 minutes\n\n**Example:**\n```\nAuthorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...\n```"
      }
    },
...
  },
  "security": [],
...
}